Don't set expected_year to 0 on /overview
We were using `expected_year == 0` to mean "There's no expected year". If it
was just internal, it would be OK, but now that we're exposing this as an API,
it starts generating confusion for the users.

With this patch, we simply set `expected_year = undefined` if it is undefined,
and that doesn't appear on the API. The only other part we use expected_year in
the code is when testing if the document is late or not (`expected_year >
year`). As `undefined > ANY_NUMBER` is false, it still works fine.