Increase year range for documents and allow consecutive years
Some countries do have fiscal years which cross two years, i.e.
do not follow the calendar year so we need to allow years such as
2013/2014 and not only full years.

Also it has become necessary to allow in some cases to add older
documents so the range is increased +/-2 years to the current year.